Refactoring results Table
    - combine logic of QueryPage-ResultData & IsolatesTable & QueryPage
    => SetResultData
    - separate services
    - split return of QueryPage into separate components for Content and DrawerControl

Ticket: #90
